requirements.txt
discord.py>=2.5.2
python-dotenv>=1.1.0
jinja2>=3.1.6
aiohttp>=3.12.14
aiohttp_jinja2>=1.6
pytest>=8.4.1
ruff>=0.12.3